final fixes

+++ user-service/src/main/java/ru/services/user/controller/UserController.java
@@ -6,28 +6,23 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
+import ru.services.user.authorization.AccessControlService;
 import ru.services.user.mapper.UserMapper;
 import ru.services.user.model.UserRequest;
 import ru.services.user.model.UserResponse;
 import ru.services.user.service.UserService;
 
 @RestController
-@RequestMapping("/users")
+@RequestMapping("/api/users/v1")
 @RequiredArgsConstructor
 public class UserController {
     private final UserService userService;
+    private final AccessControlService accessControlService;
     private final UserMapper mapper;
-
-    @Operation(summary = "Get user by ID")
-    @ApiResponses({
-        @ApiResponse(responseCode = "200", description = "User info"),
-        @ApiResponse(responseCode = "404", description = "User not found")
-    })
-    @GetMapping(value = "/{id}")
-    public UserResponse get(@PathVariable("id") Long id) {
-        return mapper.toResponse(userService.getUser(id));
-    }
 
     @Operation(summary = "Get user by username")
     @ApiResponses({
@@ -35,19 +30,17 @@ public class UserController {
             @ApiResponse(responseCode = "404", description = "User not found")
     })
     @GetMapping
-    public UserResponse get(@RequestParam String username) {
-        return mapper.toResponse(userService.getUser(username));
-    }
+    @PreAuthorize("isAuthenticated()")
+    public UserResponse get(
+            @RequestParam String username,
+            Authentication authentication) {
 
-    @Operation(summary = "Delete user by id")
-    @ApiResponses({
-            @ApiResponse(responseCode = "204", description = "User deleted"),
-            @ApiResponse(responseCode = "404", description = "User not found")
-    })
-    @DeleteMapping(value = "/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@PathVariable("id") Long id) {
-        userService.deleteUser(id);
+        String clientId = authentication.getName();
+
+        if (!accessControlService.checkUsernameMatches(clientId, username)) {
+            throw new AccessDeniedException("Access denied");
+        }
+        return mapper.toResponse(userService.getUser(username));
     }
 
     @Operation(summary = "Delete user by username")
@@ -56,8 +49,18 @@ public class UserController {
             @ApiResponse(responseCode = "404", description = "User not found")
     })
     @DeleteMapping
+    @PreAuthorize("isAuthenticated()")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@RequestParam("username") String username) {
+    public void delete(
+            @RequestParam("username") String username,
+            Authentication authentication) {
+
+        String clientId = authentication.getName();
+
+        if (!accessControlService.checkUsernameMatches(clientId, username)) {
+            throw new AccessDeniedException("Access denied");
+        }
+
         userService.deleteUser(username);
     }
 
@@ -67,9 +70,19 @@ public class UserController {
             @ApiResponse(responseCode = "409", description = "User already exists")
     })
     @PostMapping
+    @PreAuthorize("isAuthenticated()")
     @ResponseStatus(HttpStatus.CREATED)
-    public UserResponse create(@RequestBody @Valid UserRequest request,
-                               @RequestHeader("X-Request-ID") String idempotencyKey) {
+    public UserResponse create(
+            @RequestBody @Valid UserRequest request,
+            @RequestHeader("X-Request-ID") String idempotencyKey,
+            Authentication authentication) {
+
+        String clientId = authentication.getName();
+
+        if (!accessControlService.checkUsernameMatches(clientId, request.username())) {
+            throw new AccessDeniedException("Access denied");
+        }
+
         var user = userService.createUser(mapper.toUser(request), idempotencyKey);
         return mapper.toResponse(user);
     }
@@ -80,7 +93,15 @@ public class UserController {
             @ApiResponse(responseCode = "404", description = "User not found")
     })
     @PutMapping
-    public UserResponse update(@RequestBody @Valid UserRequest request) {
+    public UserResponse update(
+            @RequestBody @Valid UserRequest request,
+            Authentication authentication) {
+
+        String clientId = authentication.getName();
+
+        if (!accessControlService.checkUsernameMatches(clientId, request.username())) {
+            throw new AccessDeniedException("Access denied");
+        }
         var user = userService.updateUser(mapper.toUser(request));
         return mapper.toResponse(user);
     }
